fix more warnings

+++ src/data/vocab.cpp
@@ -12,7 +12,7 @@ Ptr<VocabBase> createVocab(const std::string& vocabPath, Ptr<Options> options, s
   return vocab ? vocab : createDefaultVocab();
 }
 
-int Vocab::loadOrCreate(const std::string& vocabPath,
+size_t Vocab::loadOrCreate(const std::string& vocabPath,
                         const std::vector<std::string>& trainPaths,
                         size_t maxSize) {
   size_t size = 0;
@@ -28,7 +28,7 @@ int Vocab::loadOrCreate(const std::string& vocabPath,
         trainPaths[0]);
 
     vImpl_ = createDefaultVocab();
-    size = vImpl_->findAndLoad(trainPaths[0], (int)maxSize);
+    size = vImpl_->findAndLoad(trainPaths[0], maxSize);
 
     if(size == 0) {
       auto newVocabPath = trainPaths[0] + vImpl_->canonicalExtension();
@@ -49,10 +49,10 @@ int Vocab::loadOrCreate(const std::string& vocabPath,
     size = load(vocabPath, maxSize);
   }
   LOG(info, "[data] Setting vocabulary size for input {} to {}", batchIndex_, size);
-  return (int)size;
+  return size;
 }
 
-int Vocab::load(const std::string& vocabPath, size_t maxSize) {
+size_t Vocab::load(const std::string& vocabPath, size_t maxSize) {
   if(!vImpl_)
     vImpl_ = createVocab(vocabPath, options_, batchIndex_);
   return vImpl_->load(vocabPath, (int)maxSize);
